+++ lib/src/deprecated/instruction.dart
@@ -1,3 +1,6 @@
+/// @deprecated This class is deprecated and will be removed in a future version.
+/// Please use the newer models in the main package instead.
+@deprecated
 class Instruction {
   Points? points;
   Annotation? annotation;
@@ -8,32 +11,41 @@ class Instruction {
   ExtraInfoJSON? extraInfoJSON;
   int? length;
 
-  Instruction(
-      {this.points,
-      this.annotation,
-      this.sign,
-      this.name,
-      this.distance,
-      this.time,
-      this.extraInfoJSON,
-      this.length});
+  /// @deprecated This constructor is deprecated and will be removed in a future version.
+  @deprecated
+  Instruction({
+    this.points,
+    this.annotation,
+    this.sign,
+    this.name,
+    this.distance,
+    this.time,
+    this.extraInfoJSON,
+    this.length,
+  });
 
+  /// @deprecated This factory is deprecated and will be removed in a future version.
+  @deprecated
   Instruction.fromJson(Map<String, dynamic> json) {
     points =
         json['points'] != null ? new Points.fromJson(json['points']) : null;
-    annotation = json['annotation'] != null
-        ? new Annotation.fromJson(json['annotation'])
-        : null;
+    annotation =
+        json['annotation'] != null
+            ? new Annotation.fromJson(json['annotation'])
+            : null;
     sign = json['sign'];
     name = json['name'];
     distance = json['distance'];
     time = json['time'];
-    extraInfoJSON = json['extraInfoJSON'] != null
-        ? new ExtraInfoJSON.fromJson(json['extraInfoJSON'])
-        : null;
+    extraInfoJSON =
+        json['extraInfoJSON'] != null
+            ? new ExtraInfoJSON.fromJson(json['extraInfoJSON'])
+            : null;
     length = json['length'];
   }
 
+  /// @deprecated This method is deprecated and will be removed in a future version.
+  @deprecated
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     final points = this.points;
@@ -63,6 +75,9 @@ class Instruction {
   }
 }
 
+/// @deprecated This class is deprecated and will be removed in a future version.
+/// Please use the newer models in the main package instead.
+@deprecated
 class Points {
   int? size;
   String? intervalString;
@@ -71,14 +86,19 @@ class Points {
   int? dimension;
   bool? empty;
 
-  Points(
-      {this.size,
-      this.intervalString,
-      this.immutable,
-      this.b3D,
-      this.dimension,
-      this.empty});
+  /// @deprecated This constructor is deprecated and will be removed in a future version.
+  @deprecated
+  Points({
+    this.size,
+    this.intervalString,
+    this.immutable,
+    this.b3D,
+    this.dimension,
+    this.empty,
+  });
 
+  /// @deprecated This factory is deprecated and will be removed in a future version.
+  @deprecated
   Points.fromJson(Map<String, dynamic> json) {
     size = json['size'];
     intervalString = json['intervalString'];
@@ -88,6 +108,8 @@ class Points {
     empty = json['empty'];
   }
 
+  /// @deprecated This method is deprecated and will be removed in a future version.
+  @deprecated
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['size'] = this.size;
@@ -105,19 +127,28 @@ class Points {
   }
 }
 
+/// @deprecated This class is deprecated and will be removed in a future version.
+/// Please use the newer models in the main package instead.
+@deprecated
 class Annotation {
   bool? empty;
   int? importance;
   String? message;
 
+  /// @deprecated This constructor is deprecated and will be removed in a future version.
+  @deprecated
   Annotation({this.empty, this.importance, this.message});
 
+  /// @deprecated This factory is deprecated and will be removed in a future version.
+  @deprecated
   Annotation.fromJson(Map<String, dynamic> json) {
     empty = json['empty'];
     importance = json['importance'];
     message = json['message'];
   }
 
+  /// @deprecated This method is deprecated and will be removed in a future version.
+  @deprecated
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['empty'] = this.empty;
@@ -132,15 +163,24 @@ class Annotation {
   }
 }
 
+/// @deprecated This class is deprecated and will be removed in a future version.
+/// Please use the newer models in the main package instead.
+@deprecated
 class ExtraInfoJSON {
   double? heading;
 
+  /// @deprecated This constructor is deprecated and will be removed in a future version.
+  @deprecated
   ExtraInfoJSON({this.heading});
 
+  /// @deprecated This factory is deprecated and will be removed in a future version.
+  @deprecated
   ExtraInfoJSON.fromJson(Map<String, dynamic> json) {
     heading = json['heading'];
   }
 
+  /// @deprecated This method is deprecated and will be removed in a future version.
+  @deprecated
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['heading'] = this.heading;